change policyserviceimpl, campaignserviceimpl, customerserviceimpl

+++ Project/demo/src/main/java/com/fptu/swp391/se1839/oemevwarrantymanagement/dto/response/SummaryItemResponse.java
@@ -13,6 +13,6 @@ import lombok.experimental.FieldDefaults;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class SummaryItemResponse {
-    long percentage;
+    double percentage;
     String message;
 }
